src/third_party/blink/web_tests/fast/html/imports/import-custom-element-abort-expected.txt

CONSOLE WARNING: line 8: document.registerElement is deprecated and will be removed in M73, around March 2019. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 for more details.
CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
CONSOLE MESSAGE: line 4: PASS unless crash.
